Removed custom searchBar and replaced it with MUI component
