+++ dashdoc-frontend/src/components/shared/tables/DataTable.tsx
@@ -8,7 +8,7 @@ import {
 import { defaultTableProps } from "@constants";
 import { CustomNoRowsOverlay } from "@styles/dataTable";
 import * as S from "@styles";
-import { ClientsPageToolbar } from "./CustomToolbars";
+import { ClientsPageToolbar } from "../../ui/CustomToolbars";
 
 type TableProps = {
   columns: GridColDef[];
